+++ src/main/java/com/tma/teamhr/repository/ITeamRepository.java
@@ -3,5 +3,5 @@ package com.tma.teamhr.repository;
 import com.tma.teamhr.model.Team;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ITeamRepository extends CrudRepository<Team, Long> {
+public interface ITeamRepository extends CrudRepository<Team, Integer> {
 }
